refactor(pidash-server): extract Discord bot and async handler
Extract Discord bot (~500 lines) into scripts/pidash-discord.ts with
setupDiscordBot() function. Extract async WebSocket handler into named
handleAsyncConnection() function.

pidash-server.ts reduced from 973 to 413 lines.
Pure mechanical extraction with zero logic changes.

Closes #581
